# src/swagger2.0.json
{
    "swagger": "2.0",
    "info": {
        "title": "API test platform API",
        "description": "This is the API for the API test platform. So you can test if your api's validate to the dutch government standards",
        "version": "1.0.0-rc1"
    },
    "basePath": "/api/v1",
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "securityDefinitions": {
        "Basic": {
            "type": "basic"
        }
    },
    "security": [
        {
            "Basic": []
        }
    ],
    "paths": {
...
        "/testsessions/{uuid}": {
            "get": {
                "operationId": "testsessions_read",
                "summary": "Session detail.",
                "description": "Return the given session's detail.",
                "parameters": [
                    {
                        "name": "If-None-Match",
                        "in": "header",
                        "description": "Perform conditional requests. This header should contain one or multiple ETag values of resources the client has cached. If the current resource ETag value is in this set, then an HTTP 304 empty body will be returned. See [MDN](https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/If-None-Match) for details.",
                        "required": false,
                        "type": "string",
                        "examples": {
                            "oneValue": {
                                "summary": "One ETag value",
                                "value": "\"79054025255fb1a26e4bc422aef54eb4\""
                            },
                            "multipleValues": {
                                "summary": "Multiple ETag values",
                                "value": "\"79054025255fb1a26e4bc422aef54eb4\", \"e4d909c290d0fb1ca068ffaddf22cbd0\""
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Session"
                        },
                        "headers": {
                            "ETag": {
                                "description": "De ETag berekend op de response body JSON. Indien twee resources exact dezelfde ETag hebben, dan zijn deze resources identiek aan elkaar. Je kan de ETag gebruiken om caching te implementeren.",
                                "type": "string"
                            },
                            "API-version": {
                                "schema": {
                                    "type": "string"
                                },
                                "description": "Geeft een specifieke API-versie aan in de context van een specifieke aanroep. Voorbeeld: 1.2.1."
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/Fout"
                        },
                        "headers": {
                            "API-version": {
                                "schema": {
                                    "type": "string"
                                },
                                "description": "Geeft een specifieke API-versie aan in de context van een specifieke aanroep. Voorbeeld: 1.2.1."
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/Fout"
                        },
                        "headers": {
                            "API-version": {
                                "schema": {
                                    "type": "string"
                                },
                                "description": "Geeft een specifieke API-versie aan in de context van een specifieke aanroep. Voorbeeld: 1.2.1."
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "$ref": "#/definitions/Fout"
                        },
                        "headers": {
                            "API-version": {
                                "schema": {
                                    "type": "string"
                                },
                                "description": "Geeft een specifieke API-versie aan in de context van een specifieke aanroep. Voorbeeld: 1.2.1."
                            }
                        }
                    },
                    "406": {
                        "description": "Not acceptable",
                        "schema": {
                            "$ref": "#/definitions/Fout"
                        },
                        "headers": {
                            "API-version": {
                                "schema": {
                                    "type": "string"
                                },
                                "description": "Geeft een specifieke API-versie aan in de context van een specifieke aanroep. Voorbeeld: 1.2.1."
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/Fout"
                        },
                        "headers": {
                            "API-version": {
                                "schema": {
                                    "type": "string"
                                },
                                "description": "Geeft een specifieke API-versie aan in de context van een specifieke aanroep. Voorbeeld: 1.2.1."
                            }
                        }
                    },
                    "410": {
                        "description": "Gone",
                        "schema": {
                            "$ref": "#/definitions/Fout"
                        },
                        "headers": {
                            "API-version": {
                                "schema": {
                                    "type": "string"
                                },
                                "description": "Geeft een specifieke API-versie aan in de context van een specifieke aanroep. Voorbeeld: 1.2.1."
                            }
                        }
                    },
                    "415": {
                        "description": "Unsupported media type",
                        "schema": {
                            "$ref": "#/definitions/Fout"
                        },
                        "headers": {
                            "API-version": {
                                "schema": {
                                    "type": "string"
                                },
                                "description": "Geeft een specifieke API-versie aan in de context van een specifieke aanroep. Voorbeeld: 1.2.1."
                            }
                        }
                    },
                    "429": {
                        "description": "Too many requests",
                        "schema": {
                            "$ref": "#/definitions/Fout"
                        },
                        "headers": {
                            "API-version": {
                                "schema": {
                                    "type": "string"
                                },
                                "description": "Geeft een specifieke API-versie aan in de context van een specifieke aanroep. Voorbeeld: 1.2.1."
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/Fout"
                        },
                        "headers": {
                            "API-version": {
                                "schema": {
                                    "type": "string"
                                },
                                "description": "Geeft een specifieke API-versie aan in de context van een specifieke aanroep. Voorbeeld: 1.2.1."
                            }
                        }
                    }
                },
                "tags": [
                    "testsessions"
                ]
            },
            "parameters": [
                {
                    "name": "uuid",
                    "in": "path",
                    "required": true,
                    "type": "string"
                }
            ]
        },
...
    },
    "definitions": {
...
        "Fout": {
            "required": [
                "code",
                "title",
                "status",
                "detail",
                "instance"
            ],
            "type": "object",
            "properties": {
                "type": {
                    "title": "Type",
                    "description": "URI referentie naar het type fout, bedoeld voor developers",
                    "type": "string"
                },
                "code": {
                    "title": "Code",
                    "description": "Systeemcode die het type fout aangeeft",
                    "type": "string",
                    "minLength": 1
                },
                "title": {
                    "title": "Title",
                    "description": "Generieke titel voor het type fout",
                    "type": "string",
                    "minLength": 1
                },
                "status": {
                    "title": "Status",
                    "description": "De HTTP status code",
                    "type": "integer"
                },
                "detail": {
                    "title": "Detail",
                    "description": "Extra informatie bij de fout, indien beschikbaar",
                    "type": "string",
                    "minLength": 1
                },
                "instance": {
                    "title": "Instance",
                    "description": "URI met referentie naar dit specifiek voorkomen van de fout. Deze kan gebruikt worden in combinatie met server logs, bijvoorbeeld.",
                    "type": "string",
                    "minLength": 1
                }
            }
        },
...
        "ExposedUrl": {
            "required": [
                "session"
            ],
            "type": "object",
            "properties": {
                "id": {
                    "title": "ID",
                    "type": "integer",
                    "readOnly": true
                },
                "subdomain": {
                    "title": "Subdomain",
                    "description": "The subdomain under which the service has been exposed",
                    "type": "string",
                    "maxLength": 200,
                    "minLength": 1,
                    "x-nullable": true
                },
                "session": {
                    "title": "Session",
                    "description": "The session to which this exposed URL belongs",
                    "type": "integer"
                },
                "vng_endpoint": {
                    "title": "Vng endpoint",
                    "type": "string",
                    "pattern": "^[^ ]*$",
                    "readOnly": true
                }
            }
        },
...
        "Session": {
            "required": [
                "session_type"
            ],
            "type": "object",
            "properties": {
                "uuid": {
                    "title": "Uuid",
                    "description": "The universally unique identifier of this session, needed to retrieve the badge",
                    "type": "string",
                    "format": "uuid",
                    "readOnly": true
                },
                "session_type": {
                    "title": "Session type",
                    "type": "string"
                },
                "started": {
                    "title": "Started at",
                    "description": "The time at which the session was started",
                    "type": "string",
                    "format": "date-time",
                    "readOnly": true
                },
                "stopped": {
                    "title": "Stopped at",
                    "description": "The time at which the session was stopped",
                    "type": "string",
                    "format": "date-time",
                    "readOnly": true
                },
                "status": {
                    "title": "Status",
                    "description": "Indicates the status of this session",
                    "type": "string",
                    "enum": [
                        "starting",
                        "running",
                        "shutting down",
                        "stopped",
                        "Error deployment"
                    ],
                    "readOnly": true
                },
                "exposedurl_set": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ExposedUrl"
                    },
                    "readOnly": true
                },
                "build_version": {
                    "title": "Build version",
                    "type": "string",
                    "minLength": 1
                },
                "sandbox": {
                    "title": "Sandbox",
                    "description": "If enabled, whenever multiple calls are made to the same path, the result of the call is overridden every time",
                    "type": "boolean"
                }
            }
        },
...
    }
}